Fix bugs in DSE.Open.Language projects
- FleschReadabilityCalculator: use floating-point division for
  words/sentences and syllables/words. Integer division was truncating
  these ratios to whole numbers, producing wildly incorrect ease and
  grade scores.
- SyllableCount: load syllable-counts.txt from AppContext.BaseDirectory
  directly. The file is copied flat to the output directory, but the
  loader was looking under a non-existent "Readability/" subfolder,
  causing FileNotFoundException at first use.
- WordFeatureCollection.Insert: validate that the feature name is not
  already present, matching the invariant enforced by Add.
- AttributeValueSerializer / WordFeatureSerializer TrySerialize: return
  false when the inner TryFormat fails instead of silently skipping the
  failed item and reporting success with truncated output.
- StanzaAnnotator.MapWord: report "attributes" rather than "features" in
  the error message when attribute parsing fails.

+++ src/DSE.Open.Language.Readability/FleschReadabilityCalculator.cs
@@ -36,8 +36,8 @@ public static class FleschReadabilityCalculator
     public static double CalculateReadingEase(int wordCount, int sentenceCount, int syllableCount)
     {
         return EaseBaseScore -
-            (EaseSentenceWeight * (wordCount / sentenceCount)) -
-            (EaseWordWeight * (syllableCount / wordCount));
+            (EaseSentenceWeight * ((double)wordCount / sentenceCount)) -
+            (EaseWordWeight * ((double)syllableCount / wordCount));
     }
 
     public static double CalculateReadingGrade(IReadOnlyCollection<Sentence> sentences)
@@ -60,8 +60,8 @@ public static class FleschReadabilityCalculator
 
     public static double CalculateReadingGrade(int wordCount, int sentenceCount, int syllableCount)
     {
-        return (GradeSentenceWeight * (wordCount / sentenceCount)) +
-            (GradeWordWeight * (syllableCount / wordCount)) +
+        return (GradeSentenceWeight * ((double)wordCount / sentenceCount)) +
+            (GradeWordWeight * ((double)syllableCount / wordCount)) +
             GradeBaseScore;
     }
 }
